feat(24.04): add ubuntu archive signing key 2018
This commit adds the Ubuntu Archive Automatic Signing Key (2018) as a
"public-key" in the chisel.yaml. The armored data was obtained by
executing the following commands on a Ubuntu 22.04 machine:

    $ gpg --keyserver keyserver.ubuntu.com --receive-keys 871920D1991BC93C
    $ gpg --armor --export 871920D1991BC93C

BREAKING CHANGE: This commit introduces the new chisel yaml format
"chisel-v2", in which ``public-keys`` (top-level) and
``archive.<name>.public-keys`` fields are introduced. This changes is
related to the addition of integrity checks in chisel. [1]

[1] https://github.com/canonical/chisel/pull/106