fix(24.04): remove Byte Order Marks from files
A few slice definition files in the ubuntu-24.04 release contains Byte
Order Mark (BOM) [1-2]. This is not needed for our purposes and ASCII
text is fine for the slice definition files at the moment. Hence, this
commit removes all the BOMs from the files in this branch.

Note: to check the status of the files, you may run

    file chisel.yaml slices/*

Refernces:
1. https://en.wikipedia.org/wiki/Byte_order_mark
2. https://stackoverflow.com/q/2223882