chore: check paths diff within same branch on push
On "push" events, specify the "base" value of dorny/paths-filter action.
This ensures that the comparison is made against the recent commit on
that particular branch before pushing the new commits.

Previously, it was left empty, resulting in the default value which is
the default branch "ubuntu-22.04". Thus pushed to the "ubuntu-24.04"
branch would be compared with "ubuntu-22.04" which is a bug. [1]

This commit fixes said bug.

[1] https://github.com/canonical/chisel-releases/actions/runs/8092571514/job/22113523205#step:3:52