chore(22.04): add pkg-deps job
This commit adds a new pkg-deps job on Pull Request. It checks in
changed files if slices of any package A has been listed in "essential"
of a slice in package B, but B does not depend on A. If found, the CI
will add a comment in the Pull Request with the findings.