fix(22.04/libnss3): libs should contain all libraries
This commit restructures the libnss3 slices. It renames the former
"libs" slice to "nss", to only include nss3-specific libraries. It also
adds a new slice "smime" to include S/MIME libraries.

Finally, the "libs" slice is added back, which includes all of the
slices that provide libraries. The rationale is that "libs" slices
should contain all libraries coming from the package.

Note that the ca-certificates-java slice has also been updated to
reflect this change.

Resolves #381