fix(25.10): update openssh slices for openssh 10.0
This updates the openssh slices for the removal of:

- /usr/bin/slogin
- /usr/lib/openssh/agent-launch

It also adds integration tests.

While writing the integration tests, I noticed that ssh-copy-id and
ssh-argv0 are actually shell scripts, so I separated them into their own
slices and made a scripts slice for them and added their appropriate
dependencies. These dependencies are notably absent from the actual
openssh-client package, which may be a packaging bug.
The individual slices are useful for distinguishing which slices are
necessary for which script.

We may want to backport the scripts slice to earlier releases, but I'm
unsure how best to do that without breaking compatibility for them.

Signed-off-by: Alex Lowe <alex.lowe@canonical.com>