fix: chisel path conflic between with dbg packages
This commit removes the dotnet-sdk-dbg-8.0,
aspnetcore-runtime-dbg-8.0, dotnet-runtime-dbg-8.0 slice
definition files as well as the dotnet-sdk-8.0_dev,
aspnetcore-runtime-8.0_dev, dotnet-runtime-8.0_dev slice
definitions, because it is hard to create maintainable
slice definition files.

The dbg packages contain *.pdb symbol files which are
located right next to the files they provide symbols for.
This makes it impossible to use glob patterns, because
chisel can not staticly (without downloading the packages
file contents) check for file conflicts.

I may revisit the idea of creating slices for the dbg
packages in the future.

Signed-off-by: Dominik Viererbe <dominik.viererbe@canonical.com>