refactor: aspnetcore tests
This commit increases the size of the mocked /dev/urandom and
/dev/random files in the chroot. The test hangs indefinetly if
these tests are run with .NET 10, because it tries to read more
than 500 characters from these files and then is stuck in an
infinte loop, because it repeatedly reads 0 characters at the
end of file, which will never fill the read buffer to its target size.

The .NET 8 & 9 tests technically do not need this change, but I want
to keep the tests as consistent as possible between the major versions
when I later add the .NET 10 slices.